# Improving array copy performance in `StylusPoint.CopyAdditionalData` by using Span