@halgab 2023-11-10
What about `_additionalValues.AsSpan().ToArray()`?