@halgab 2023-11-10
I think there is a mistake in your benchmark. It should be

```csharp
    [Benchmark]
    public void CopyAdditionalDataSpanToArray()
    {
        if (null != _additionalValues)
        {
            _additionalValues = _additionalValues.AsSpan().ToArray();
        }
    }
```

or 

```csharp
    [Benchmark]
    public void CopyAdditionalDataSpanToArray()
    {
        _additionalValues = _additionalValues?.AsSpan().ToArray();
    }
```