@IAmTheCShark 2023-11-19
The fact that `AsSpan().ToArray()` is faster probably boils down to the fact that `Array.Clone` uses `Object.MemberWiseClone`
Maybe this is something that should be a proposed as an improvement for the Clone method. I dont think changing the code like this in the WPF repo is a good idea in that kind of regard.